Inventory Component / Weapon Pickup Logic
Created an Inventory Component which stores an array for which weapons the player has.

The Component has two functions (so far), to either Find the Weapon in the inventory or Add it

The Add Function will add the new Weapon structure to an index in the inventory array, however if the player already has the weapon it will give the player max ammo.

If the player has the weapon and is at max ammo they cannot pick up the weapon.