Animation and Weapon attachment
Added animation to the character to simulate movement with a rifle, changes were made in the Animation Blueprint;

Modified the weapon pickup to spawn the weapon actor and attach it to the character as a child actor

Enabled the crouch mechanic on the character movement and gave it proper animations/blend spaces.

Note: Animations were not created by me, License will have to be checked.